sql: add degree trig and hyperbolic trig builtins
Previously, only radian trig builtins were supported.
All trig builtins were also previously untested.
Adds the missing trig builtins and adds tests for all trig builtins

Release note: None

sql: break out math builtins to separate file from builtins

Previously, the math builtins were all in the master builtins file.
The builtins file was too large.
This patch moves the math builtins to a separate file to make builtins more manageable.

Release note: None
