+++ pkg/sql/sem/tree/testdata/eval/trig
@@ -1,0 +1,111 @@
+# Trig functions
+
+eval
+cos(0)
+----
+1.0
+
+eval
+sin(1.5707963267948966)
+----
+1.0
+
+eval
+tan(0)
+----
+0.0
+
+eval
+cot(0.7853981633974483)
+----
+1.0
+
+eval
+acos(0)
+----
+1.5707963267948966
+
+eval
+asin(1.0)
+----
+1.5707963267948966
+
+eval
+atan(1.0)
+----
+0.7853981633974483
+
+eval
+atan2(1.0, 1.0)
+----
+0.7853981633974483
+
+eval
+cosd(0)
+----
+1.0
+
+eval
+sind(90)
+----
+1.0
+
+eval
+tand(45)
+----
+1.0
+
+eval
+cotd(45)
+----
+1.0
+
+eval
+acosd(0)
+----
+90.0
+
+eval
+asind(1.0)
+----
+90.0
+
+eval
+atand(1.0)
+----
+45.0
+
+eval
+atan2d(1.0, 1.0)
+----
+45.0
+
+eval
+sinh(0)
+----
+0.0
+
+eval
+cosh(0)
+----
+1.0
+
+eval
+tanh(0)
+----
+0.0
+
+eval
+asinh(0)
+----
+0.0
+
+eval
+acosh(1)
+----
+0.0
+
+eval
+atanh(0)
+----
+0.0
